feedback patch // moved logic to lib index file // added hoek for err assertion // updated license // updated gitignore file (taken from hapi repo)
