+++ .gitignore
@@ -1,7 +1,11 @@
 .idea
 *.iml
 npm-debug.log
+dump.rdb
 node_modules
+results.tap
+results.xml
+config.json
 .DS_Store
 */.DS_Store
 */*/.DS_Store
